requirements.txt
# Core framework
Flask>=3.1.1,<4.0
Flask-Limiter>=3.3.1,<4.0

# OpenAI & LangChain
openai>=1.78.1,<2.0
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.16

# Vector store
faiss-cpu>=1.11.0,<2.0

# Docs & slides
python-docx>=1.1.2
python-pptx>=0.6.23
PyMuPDF>=1.25.5

# Data & plots
pandas>=2.2.1,<3.0
matplotlib>=3.8.2,<4.0

# Storage & I/O
fsspec[http]>=2023.1,<2024.10
requests>=2.32.2,<3.0

# Hugging Face & utilities
huggingface-hub>=0.30.2,<1.0
tqdm>=4.66.3,<5.0

# Validation & serialization
pydantic>=2.10,<3.0
protobuf>=4.21.6,<5.0

# Auth & OS helpers
PyJWT[crypto]>=2.5.0,<3.0
python-magic-bin; platform_system == "Windows"
